chore: consolidate all the types in one place
