+++ packages/ipfs-core/src/components/root.js
@@ -7,18 +7,17 @@ const createGetAPI = require('./get')
 const createLsAPI = require('./ls')
 
 /**
+ * @typedef {AddAllContext & CatContext & GetContext & ListContext } Context
+ * @typedef {import('./add-all').Context} AddAllContext
+ * @typedef {import('./cat').Context} CatContext
+ * @typedef {import('./get').Context} GetContext
+ * @typedef {import('./ls').Context} ListContext
  * @typedef {import('ipfs-interface').RootAPI} RootAPI
  * @implements {RootAPI}
  */
 class Root {
   /**
-   * @param {Object} config
-   * @param {import('.').Block} config.block
-   * @param {import('.').Pin} config.pin
-   * @param {import('.').GCLock} config.gcLock
-   * @param {import('.').Preload} config.preload
-   * @param {import('.').IPLD} config.ipld
-   * @param {import('ipfs-interface/src/root').ShardingOptions} [config.options]
+   * @param {Context} context
    */
   constructor ({ preload, gcLock, pin, block, ipld, options }) {
     const addAll = createAddAllAPI({
