+++ packages/ipfs-core/src/utils/service.js
@@ -212,7 +212,7 @@ module.exports = Service
 
 /**
  * @template T
- * @typedef {import('../interface/basic').Await<T>} Await
+ * @typedef {import('ipfs-interface/src/basic').Await<T>} Await
  */
 /**
  * @template {(options:any) => any} T
