+++ packages/ipfs-interface/src/basic.ts
@@ -1,4 +1,20 @@
 /**
+ * Represents a value that you can await on, which is either value or a promise
+ * of one.
+ */
+export type Await<T> =
+  | T
+  | Promise<T>
+
+/**
+ * Represents an iterable that can be used in `for await` loops, that is either
+ * iterable or an async iterable.
+ */
+export type AwaitIterable<T> =
+  | Iterable<T>
+  | AsyncIterable<T>
+
+/**
  * Common options across all cancellable requests.
  */
 export interface AbortOptions {
@@ -14,18 +30,10 @@ export interface AbortOptions {
   timeout?: number
 }
 
-/**
- * Represents a value that you can await on, which is either value or a promise
- * of one.
- */
-export type Await<T> =
-  | T
-  | Promise<T>
-
-/**
- * Represents an iterable that can be used in `for await` loops, that is either
- * iterable or an async iterable.
- */
-export type AwaitIterable<T> =
-  | Iterable<T>
-  | AsyncIterable<T>
+export type ToJSON =
+  | null
+  | string
+  | number
+  | boolean
+  | ToJSON[]
+  | { toJSON?: () => ToJSON } & { [key: string]: ToJSON }
