+++ packages/ipfs-interface/src/files.ts
@@ -170,42 +170,3 @@ type HRTime = [number, number]
 
 // It's just a named type alias, but it better captures intent.
 export type Mode = number
-
-/**
- * @typedef {import('../format-mtime').MTime} MTime
- * @typedef {import('../format-mode').Mode} Mode
- * @typedef {Object} Directory
- * @property {string} path
- * @property {Mode} [mode]
- * @property {MTime} [mtime]
- * @property {undefined} [content]
- *
- * @typedef {Object} FileInput
- * @property {string} [path]
- * @property {ToContent} [content]
- * @property {number | string} [mode]
- * @property {UnixTime} [mtime]
- *
- * @typedef {Date | MTime | HRTime} UnixTime
- *
- * Time representation as tuple of two integers, as per the output of
- * [`process.hrtime()`](https://nodejs.org/dist/latest/docs/api/process.html#process_process_hrtime_time).
- * @typedef {[number, number]} HRTime
- *
- * @typedef {string|InstanceType<typeof window.String>|ArrayBufferView|ArrayBuffer|Blob|Iterable<Uint8Array> | AsyncIterable<Uint8Array> | ReadableStream<Uint8Array>} ToContent
- * @typedef {ToContent|FileInput} ToFile
- * @typedef {Iterable<ToFile> | AsyncIterable<ToFile> | ReadableStream<ToFile>} Source
- */
-/**
- * @template {AsyncIterable<Uint8Array>|Blob} Content
- * @typedef {Object} File
- * @property {string} path
- * @property {Mode} [mode]
- * @property {MTime} [mtime]
- * @property {Content} [content]
- */
-
-/**
- * @template {AsyncIterable<Uint8Array>|Blob} Content
- * @typedef {File<Content>|Directory} Entry
- */
